+++ uFood.API/Controllers/ValuesController.cs
@@ -1,5 +1,5 @@
-﻿using System.Collections.Generic;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
+using System.Collections.Generic;
 
 namespace uFood.API.Controllers
 {
@@ -11,7 +11,7 @@ namespace uFood.API.Controllers
 		[HttpGet]
 		public ActionResult<IEnumerable<string>> Get()
 		{
-			return new string[] {"value1", "value2"};
+			return new[] {"value1", "value2"};
 		}
 
 		// GET api/values/5
