+++ uFood.Infrastructure.LichtBild/Helper/PhotographyHelper.cs
@@ -1,19 +1,17 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using System.Collections.Generic;
 using uFood.Infrastructure.LichtBild.Model;
 using uFood.Infrastructure.Models.Environment;
 
 namespace uFood.Infrastructure.LichtBild.Helper
 {
-    public class PhotographyHelper
-    {
-        public List<Photography> GetPhotographiesByPosition(Position position, double searchRadius)
-        {
-            List<Photography> photographies = new List<Photography>();
+	public class PhotographyHelper
+	{
+		public IEnumerable<Photography> GetPhotographiesByPosition(Position position, double searchRadius)
+		{
+			var photographies = new List<Photography>();
 
 
-            return photographies;
-        }
-    }
+			return photographies;
+		}
+	}
 }
