+++ uFood.Infrastructure.LichtBild/Model/Photography.cs
@@ -1,13 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace uFood.Infrastructure.LichtBild.Model
+﻿namespace uFood.Infrastructure.LichtBild.Model
 {
-    public class Photography
-    {
-        public string ID { get; set; }
+	public class Photography
+	{
+		public string ID { get; set; }
 
-        public string ForeginID { get; set; }
-    }
+		public string ForeignID { get; set; }
+	}
 }
