+++ uFood.Infrastructure.LichtBild/ModelExtensions/PhotographyExtension.cs
@@ -1,11 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace uFood.Infrastructure.LichtBild.ModelExtensions
+﻿namespace uFood.Infrastructure.LichtBild.ModelExtensions
 {
-    public class PhotographyExtension
-    {
-
-    }
+	public class PhotographyExtension
+	{
+	}
 }
