+++ uFood.Infrastructure.OpenDataHub/Model/Gastronomy.cs
@@ -1,17 +1,14 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using System.Collections.Generic;
 using uFood.Infrastructure.Models.Food;
 
 namespace uFood.Infrastructure.OpenDataHub.Model
 {
-    public class Gastronomy
-    {
-        public string ID { get; set; }
+	public class Gastronomy
+	{
+		public string ID { get; set; }
 
-        public string ForeginID { get; set; }
+		public string ForeignID { get; set; }
 
-        public IEnumerable<Dish> Dishes { get; set; }
-
-    }
+		public IEnumerable<Dish> Dishes { get; set; }
+	}
 }
