+++ uFood.Infrastructure.OpenDataHub/ModelExtensions/GastronomyExtension.cs
@@ -1,19 +1,17 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using System.Collections.Generic;
 using uFood.Infrastructure.Models.Food;
 using uFood.Infrastructure.OpenDataHub.Model;
 
 namespace uFood.Infrastructure.OpenDataHub.ModelExtensions
 {
-    public static class GastronomyExtension
-    {
-        public static  List<Dish> GetDishesByNutrient(this Gastronomy gastronomy, Nutrient nutrient)
-        {
-            List<Dish> disesh = new List<Dish>();
+	public static class GastronomyExtension
+	{
+		public static IEnumerable<Dish> GetDishesByNutrient(this Gastronomy gastronomy, Nutrient nutrient)
+		{
+			var dishes = new List<Dish>();
 
 
-            return disesh;
-        }
-    }
+			return dishes;
+		}
+	}
 }
