+++ uFood.Infrastructure/Models/Environment/Farmer.cs
@@ -1,17 +1,15 @@
 ﻿namespace uFood.Infrastructure.Models.Environment
 {
-    public class Farmer
-    {
-        public string ID { get; set; }
+	public class Farmer
+	{
+		public string ID { get; set; }
 
-        public string Name { get; set; }
+		public string Name { get; set; }
 
-        public string Address { get; set; }
+		public string Address { get; set; }
 
-        public string Contact { get; set; }
+		public string Contact { get; set; }
 
-        public Position Position { get; set; }
-
-    }
-
+		public Position Position { get; set; }
+	}
 }
