+++ uFood.Infrastructure/Models/Environment/NutrientPosition.cs
@@ -1,16 +1,13 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using uFood.Infrastructure.Models.Food;
+﻿using uFood.Infrastructure.Models.Food;
 
 namespace uFood.Infrastructure.Models.Environment
 {
-    public class NutrientPosition
-    {
-        public string ID { get; set; }
+	public class NutrientPosition
+	{
+		public string ID { get; set; }
 
-        public Nutrient Nutrient { get; set; }
+		public Nutrient Nutrient { get; set; }
 
-        public Farmer Farmer { get; set; }
-    }
+		public Farmer Farmer { get; set; }
+	}
 }
