+++ uFood.Infrastructure/Models/Environment/Position.cs
@@ -1,15 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace uFood.Infrastructure.Models.Environment
+﻿namespace uFood.Infrastructure.Models.Environment
 {
-    public class Position
-    {
-        public double Latitude { get; set; }
+	public class Position
+	{
+		public double Latitude { get; set; }
 
-        public double Longitude { get; set; }
+		public double Longitude { get; set; }
 
-        public double Altitude { get; set; }
-    }
+		public double Altitude { get; set; }
+	}
 }
