+++ uFood.Infrastructure/Models/Food/Dish.cs
@@ -1,17 +1,13 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace uFood.Infrastructure.Models.Food
+﻿namespace uFood.Infrastructure.Models.Food
 {
-    public class Dish
-    {
-        public string ID { get; set; }
+	public class Dish
+	{
+		public string ID { get; set; }
 
-        public string Name { get; set; }
+		public string Name { get; set; }
 
-        public string Description { get; set; }
+		public string Description { get; set; }
 
-        public Recipe Recipe { get; set; }
-    }
+		public Recipe Recipe { get; set; }
+	}
 }
