+++ uFood.Infrastructure/Models/Food/Ingredient.cs
@@ -1,13 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace uFood.Infrastructure.Models.Food
+﻿namespace uFood.Infrastructure.Models.Food
 {
-    public class Ingredient
-    {
-        public double Quantity { get; set; }
+	public class Ingredient
+	{
+		public double Quantity { get; set; }
 
-        public Nutrient Nutrient { get; set; }
-    }
+		public Nutrient Nutrient { get; set; }
+	}
 }
