+++ uFood.Infrastructure/Models/Food/Nutrient.cs
@@ -1,15 +1,11 @@
-﻿using System;
-
-namespace uFood.Infrastructure.Models.Food
+﻿namespace uFood.Infrastructure.Models.Food
 {
-    public class Nutrient
-    {
-        public string ID { get; set; }
+	public class Nutrient
+	{
+		public string ID { get; set; }
 
-        public string Name { get; set; }
+		public string Name { get; set; }
 
-        public string Description { get; set; }
-
-    }
-
+		public string Description { get; set; }
+	}
 }
