+++ uFood.Infrastructure/Models/Food/Recipe.cs
@@ -1,16 +1,13 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using System.Collections.Generic;
 
 namespace uFood.Infrastructure.Models.Food
 {
-    public class Recipe
-    {
-        public string ID { get; set; }
+	public class Recipe
+	{
+		public string ID { get; set; }
 
-        public string Name { get; set; }
+		public string Name { get; set; }
 
-        public IEnumerable<Ingredient> Ingredients { get; set; }
-
-    }
+		public IEnumerable<Ingredient> Ingredients { get; set; }
+	}
 }
